added tokens_utils.c and struct optimization
