+++ src/parser_utils.c
@@ -2,7 +2,7 @@
 //helper functions for the parser
 void	handle_input_redirection(t_cmd *current_cmd, t_token **tokens)
 {
-	if ((*tokens)->next || (*tokens)->next->type != TOKEN_WORD)
+	if (!(*tokens)->next || (*tokens)->next->type != TOKEN_WORD)
 		exit_parsing(NULL, "syntax error near unexpected token `newline'");
 	if (current_cmd->infile)
 		free(current_cmd->infile);
@@ -25,7 +25,7 @@ void	handle_word(t_cmd *current_cmd, t_token *tokens)
 	add_arg_to_cmd(current_cmd, tokens->value);
 }
 
-void	handle_pipe(t_cmd **current_cmd)
+void	handle_pipe(t_cmd **current_cmd, t_token **tokens)
 {
 	if (!(*tokens)->next || (*tokens)->next->type == TOKEN_PIPE)
 		exit_parsing(NULL, "syntax error near unexpected token `|'");
