+++ src/test_main.c
@@ -1,4 +1,4 @@
-nclude "minishell.h"
+#include "minishell.h"
 
 int	main(int argc, char **argv)
 {
